+++ public/index.php
@@ -1,8 +1,5 @@
 <?php
 
-require dirname(__DIR__)."/vendor/autoload.php";
+require_once dirname(__DIR__)."/vendor/autoload.php";
 
-use App\Application;
-
-$app = new Application;
-$app->hello();
+require dirname(__DIR__)."/config/database.php";