+++ src/Application.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace App;
-
-class Application{
-    public function hello():void{
-        echo "hello";
-    }
-}